Refactors to improve consistency
