+++ src/api/java/baritone/api/schematic/AbstractSchematic.java
@@ -17,8 +17,6 @@
 
 package baritone.api.schematic;
 
-import baritone.api.utils.ISchematic;
-
 public abstract class AbstractSchematic implements ISchematic {
 
     protected int x;
