+++ src/api/java/baritone/api/schematic/ShellSchematic.java
@@ -17,7 +17,6 @@
 
 package baritone.api.schematic;
 
-import baritone.api.utils.ISchematic;
 import net.minecraft.block.state.IBlockState;
 
 public class ShellSchematic extends MaskSchematic {
