+++ src/api/java/baritone/api/utils/accessor/IItemStack.java
@@ -15,7 +15,7 @@
  * along with Baritone.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package baritone.api.accessor;
+package baritone.api.utils.accessor;
 
 public interface IItemStack {
