+++ src/main/java/baritone/utils/schematic/schematica/SchematicAdapter.java
@@ -17,7 +17,7 @@
 
 package baritone.utils.schematic.schematica;
 
-import baritone.api.utils.ISchematic;
+import baritone.api.schematic.ISchematic;
 import com.github.lunatrius.schematica.client.world.SchematicWorld;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.util.math.BlockPos;
